ai-service/requirements.txt
Flask==3.0.2
google-generativeai==0.4.0
python-dotenv==1.0.1
requests==2.31.0